Fix some bug so that the grasp can work properly.And save the context switch time to the file switchtime.
When counting out_time and in_time, it originally using int type .This will cause an error when dividing by 100.The number after the decimal point will be eliminate. I change the int type to float type.
I save the result of in_time-out_time to the file named switchtime to show the cost of context switch time.